Bug#18184868: CANNOT USE BOOTSTRAP VARIABLES IN -MASTER.OPT FILE
Issue:
------
Non-dynamic variables such as innodb_page_size and
innodb_undo_tablespaces cannot be set inside master.opt files because
the server throws an error during resizing, if the value passed differs
from the one that was used to initialize the data directory.
This is because master.opt files were used only to start the server with a
particular configuration, and not for initializing the server with that
configuration. The solution is to reinitialize the data directory that MTR
started with when such variables are set.

Fix:
----
When non-dynamic variables are set in the master.opt file, they must be
prefixed with "--bootstrap" so that MTR knows that it must delete, and
reinitialize the data directory with the values that have been passed.
Then, MTR starts the server with the values that have been set in the file.

Reviewed-by: Pavan Naik <pavan.naik@oracle.com>
Reviewed-by: Bjorn Munch <bjorn.munch@oracle.com>
RB: 14294