BUG#25170698: GR+MTS: ER1755 CAN BE IMPROVED FOR MTS LC -> MTS DATABASE
Group Replication only does support logical clock parallel applier.
Despite that we may replicate to a external server from the group,
and that external slave may be configured with database parallel
applier.
When this topology is in place, the external server would refuse to
replicate the View_change_log_events from the group, since those do
not include database context. Applier was, preemptively stopping
instead of continuing and cause possible inconsistencies.
Though View_change_log_events are NOP events when are outside of a
group, they are just written to binary log, so we are safe to apply
them and continue applying replication events.

To avoid the database parallel applier stop when it handles a
View_change_log_event, we mark this event type as IGNORABLE, which
allows database parallel applier to execute it despite not having
database context.
Also, marking View_change_log_events and
Transaction_context_log_event as IGNORABLE avoids:
 1) applier binary log decoding when reading 5.7+ binary/relay log
    files with 5.6 client tools;
 2) replication stop while replicating from a GR group to a 5.6
    server, directly or through a 5.7 slave between the group and
    the 5.6 server.