Fix for failure of dd_bootstrap.test introduced by WL#883 (CTEs).
In dd_bootstrap.test, we set up a fresh data directory with
mysqld --initialize, stop mysqld, remove the InnoDB tablespace
file of the mysql.version table, restart
mysqld with no option. mysqld is expected to:
- notice the absence of the tablespace
- consider it an upgrade from an older MySQL version
- try to create 'version'
- and stop with error, as 'version' is actually left in the DD.

Here's how it stops with error: it tries to create mysql.version:
- the function is ha_create_table(), it calls the next:
- open_table_from_share() creates a TABLE object for mysql.version,
and a dummy ha_innobase for this TABLE
- when trying to insert mysql.version into the data dictionary,
ha_innobase::create() fails with a duplicate key error, as mysql.version
"already exists" in the DD. The failure inside create() emits:
 Table '`mysql`.`version`' already exists
in the log, thanks to a my_error() around
ha_innodb.cc:10755. "duplicate key" error code is returned upwards,
after a first translation in ha_innodb.cc:
	case DB_DUPLICATE_KEY:
		/* Be cautious with returning this error, since
		mysql could re-enter the storage layer to get
		duplicated key info, the operation requires a
		valid table handle and/or transaction information,
		which might not always be available in the error
		handling stage. */
		return(HA_ERR_FOUND_DUPP_KEY);
(pay attention to the comment, it's relevant).
ha_create_table() then reports this error by
calling handler::print_error() on the still-dummy 'handler' of mysql.version,
resulting in this printout (before the WL):
 Can't write; duplicate key in table 'version'
Note that this line is absurd: the duplicate key happened when trying
to insert a row in *another* DD table than 'version'.
handler::print_error() then has this:
  case HA_ERR_FOUND_DUPP_KEY:
  {
    uint key_nr= table ? get_dup_key(error) : -1;
Before the WL, handler->table==nullptr, so get_dup_key() wasn't called.
Note that it was an inconsistent state of objects:
table->file->table was nullptr even though this 'file' logically
has a 'table' as it's table->file.
The WL introduced an "else" branch in open_table_from_share():
  else if (outparam->file) // if db_stat!=0, ha_open() set those pointers:
    outparam->file->change_table_ptr(outparam, share);
which is taken for mysql.version. It sets table->file->table to 'table',
which is a logical thing. Alas, when the creation later fails,
we go into handler::print_error() where 'table' is now non-null,
so get_dup_key() is called, but the still-dummy ha_innobase isn't in a
state ready to support get_dup_key() (in detail: get_dup_key calls info()
which calls info_low() which calls update_thd() which crashes as
it has no 'prebuilt').

Fix: make innodb not return "dup key" on creation errors, as the "handler"
isn't ready for printing "dup key" and it's a bit meaningless to report
"dup key" to the user. Such idea is already used and explained at
ha_innodb.cc:14039 (introduced by fix for MySQL bug 45961).