Bug #25095582: OPT_TRACE.BUGS_NO_PROT_ALL FAILS ON PB2 WITH                RESULT CONTENT MISMATCH
Adding analyze table to the tests added as part of
Bug #23227428 and Bug #23259872.

For Bug #23227428, there is no need for a full trace
output. We are looking for a specific string to make sure
that the range optimizer is re-run with the new index.

This is to avoid the sporadic failures seen on pb2.