Bug#25213029: TRUNK FAILS TO BUILD; TABLE.CC(7279): ERROR C2672:               'STD::MIN': NO MATCHING OVERL
MAX_INDEXES is defined as 64U by default, so it's an unsigned integer.
If you specify a custom value for it when you run cmake, and you leave
out the U suffix, it will be defined as a signed integer instead. This
confuses the compiler when it is used in a context where an unsigned
integer is expected, and the build fails.

Fix: Automatically add the U suffix to MAX_INDEXES if it is specified
without a suffix.

Additional fixes:

- raise an error if an invalid value is provided (not an unsigned integer)

- set MAX_INDEXES to 64 if a lower value is provided, as that is what
  the documentation says it should do

Change-Id: I05a9ec10361e5ec3e2c245572f860c1acbd4699e