Bug #11757421: SEND START OF NODE START COMMAND IGNORED IN RESTART
Post push fix

The patch caused the failure of 3 test cases: Bug16772,
ArbitrationWithApiNodeFailure, and EventOperationApplier_NS

The failures were due to the change in behavior caused by the
patch. The data node restarts were expected to fail but the
failures were silently suppressed previously. The tests are now
updated to expect a failure in these cases.