Bug#25056856 NDB_RPL_DIST_PRIV FAILS
The failures seem to be due to strict sql_mode checks for timestamp values.
Relaxing sql_mode while running the stored procedures. This is similar to
what is done when the tables are created during installation process.
The sql_mode is here however set to not alllow running procedures if ndb
is not configured (no default substition of storage engine).