Bug#25228584 ST_ASBINARY DOESN'T ACCEPT SECOND PARAMETER
This is a regression introduced by the fix for bug #25027963.

Problem: ST_ASBINARY doesn't accept it's second parameter.

During the refactoring done in bug #25027963, ST_ASBINARY was defined as
a function that only accepts one parameter instead of a function that
takes one or two parameters. This made ST_ASBINARY and ST_ASWKB, that
are aliases to the same function implementation, behave
differently. However, since only ST_ASWKB was tested with two
parameters, this was not discovered during refactoring.

Fix: Define ST_ASBINARY as a function that takes one or two parameters
and add more tests of ST_ASBINARY.

Change-Id: I0b0ed078a92060a8bb81c6eab3eda5839e38b462