Bug #24810564: INCORRECT BEHAVIOR OF UPDATE TABLE IN A                STORED PROCEDURE
Issue:
======
This problem occurs under the following conditions:
1) Multi-table update statement with a derived table that
   will be optimized away.
2) The update statement is part of a procedure and has a
   specific DEFINER.
3) More than one call of the procedure.

Before checking privileges on a table, the user is granted
select authorization on derived tables. When the procedure
is called for the first time, this is done during the
prepare phase by the optimizer while resolving the derived
table (see TABLE_LIST::resolve_derived). In this case that
derived table is merged into the outer query by the
optimizer.

On the second execution, multi-table update doesn't
distinguish between derived tables and regular tables and
unsets the required grant privileges already assigned to
the derived table's object. This happens when a merged
derived table is part of a materialized derived table.
Optimizer ignores the derived table object since it has
been merged into the outer query. This will result in
incorrect behavior when checking for permissions on the
derived table.

Solution:
=========
For multi-table update, the precheck phase doesn't
distinguish between regular tables and derived tables that
are part of the subqueries.

The fix is to make sure that derived tables are handled
differently. While setting permissions, currently
next_local is used to traverse the table list. This results
in skipping for some derived tables. This has now been
changed next_global. And due to this, it is also not
neccessary to handle the subqueries in the table list
separately.

Similar problem exists in the precheck code of CREATE VIEW.
This has also been modified on 8.0.