BUG#25209109 - NOT ALL TRACE MESSAGES ARE ENABLED IN GCS WITH -DWITH_TRACE_OUTPUT=1
Description:
-------------
GR lost the ability to set on GCS the debug and trace flags,
that are so much needed for developers.

Suggested Fix:
---------------
When of the unification of the CMake files, on mysql 5.7.17,
the switches that allowed GCS debug and trace were not correctly
implemented, thus causing trace and debug messages not to be
displayed at all.

This patch reintroduces those build switches and also corrects
one file that has never been built with debug directives.

- WITH_GCS_LOG_TRACE=1, will enable trace and debug messages
- WITH_GCS_LOG_DEBUG=1, will enable debug messages

This patch also removes the automatic feature that would
enable debug messages on Debug builds on GCS.