BUG#24822686: MTS WORKER IS NOT REPORTING ALL ERRORS THAT STS APPLIER               DOES
Problem
-------

The MTS slave applier is stopping because of an (injected) error, but is
reporting no useful information for troubleshooting about was has
happened.

Analysis
--------

All issues detected when applying binary log events are reported using
rli->report().

When an issue is not reported by the log event being applied, there is a
workaround to report the issue using rli->report().

The workaround was missing in the MTS applier.

Fix
---

Added the same workaround present on STS applier to the MTS applier.

Also, this patch also refactored many test cases that were expecting
the generic MTS error code from the slave to expect the specific issue
error code instead.