BUG#22904496: TEST FAILURE DUE TO STOPPED SLAVE COORDINATOR AND WORKER               THREADS
Problem
-------

The rpl_parallel_switch_sequential test case is sporadically failing
because of an unexpected error message in the slave server error log
generated by the MTS slave applier.

Analysis
--------

The test case has some non-deterministic activity that can
enable/disable MTS, generate some workload on the master and stop/start
the slave threads.

As the test case is not doing any synchronization with the master before
issuing STOP SLAVE, there is a chance that the slave applier be in MTS
mode and in a condition that it will throw the reported error message.

Fix
---

Just uncommented a suppression for the error message that was already
present on the test case.