Bug#25166648::ADD CHECK_PLUGIN_DIR.INC IN TESTS WHICH NEED MULTIPLE PLUGIN INSTALLATIONS
Problem Description:
As we know in mysql source, all the plugins might not be in the same
location. Some MTR tests would require multiple plugins, so it is desirable
that there is a check to see that, if all the plugins are in a directory,
then the test should run, else it needs to be skipped.

Fix: Added check_plugin_dir.inc in the MTR tests which requires multiple
plugin installations.

Reviewed by:
Vikram G.G <vikram.gg@oracle.com>
Srikanth B.R <srikanth.b.r@oracle.com>