Bug #11757421: SEND START OF NODE START COMMAND IGNORED IN RESTART
Post push fix

The ArbitrationWithApiNodeFailure test case failed when run with
a cluster consisting of 4 data nodes. The problem was that the
test case was designed for a cluster consisting of only 1
nodegroup.

With multiple nodegroups, a single node failure is not enough to
force arbitration. Since the single node which failed does not
form a viable community by itself, arbitration is skipped. The
fix is to simply skip the test case for clusters with more than
1 nodegroup.