Bug #24666839 INNODB'S ESTIMATE FOR ROWS-IN-TABLE CAN RETURN ROWS=1 DUE TO A RACE CONDITION
PROBLEM

While updating stats there could be a possible race condition while the
table stats are reset and then table is overwriiten with new stats might
leading a thread to read estimate rows as 1(wrong value after reset)
which would let optimizer choose a wrong plan for execution.

FIX

Simple fix from rb#14204 in trunk is backported to 5.7 that would reset
the stats only when the index is corrupted and the stats would be
directly overwritten otherwise.

Reviewed-by: Jimmy Yang <Jimmy.Yang@oracle.com>
RB: 14929