Post Merge Fix
The ndb_rpl_innodb_trans test started failing after the recent merge.
The failure was due to the fix for Bug#24577194 : LOAD DATA ERROR
1262 BEHAVIOUR DIFFERENT BETWEEN 5.6 AND 5.7 UNDER STRICT SQLMOD

It seems like the test case was incorrectly updated when the
rpl_loaddata.dat was changed previously but the issue only
showed up as a warning before the above mentioned fix.

The fix is to change the tables created to match the data
being loaded thus suppressing the error (and warning)