mtr: mem is pure flag / add /run/shm
Since 716621db3f3055781e24f561325cec6eac181717, opt_mem is
a pure flag. However it is assigned to the value of $MTR_MEM.

This leads to conflicts when MTR_MEM=/xxx/yy ./mtr --mem is
invoked. Here the --mem option overrided opt_mem leaving the
default path to be chosen.

This change makes when MTR_MEM set, opt_mem, the flag, is also
set. Both the environment and flag can no be set without conflicting.

/run/shm is added to the beginning of the path as its now used
in a few distros and is conceptually better than /dev for its purpose.

Signed-off-by: Daniel Black <daniel.black@au.ibm.com>