BUG#25182306 : --ONLY-BIG-TEST OPTION TO MTR TO RUN ONLY BIG TESTS
Description :
=============
MTR has the option --big-test to run tests that contain the
have_big.inc directive. In this case, MTR will run all ordinary
tests and the big tests (which are normally skipped). It would
be desirable both for developers and PB2 to have an option where
ordinary tests are skipped and only the big tests are run.

Fix :
=====
Introduced a new MTR option 'only-big-test' which when passed
will run only the big tests and the normal tests(non big tests)
will be skipped. If both --big-test and --only-big-test options
are passed then MTR will disable the '--only-big-test' option.

Reviewed-by: Deepa Dixit <deepa.dixit@oracle.com>
Reviewed-by: Srikanth B R <srikanth.b.r@oracle.com>
RB: 14883