Bug#25229315: ASSERTION `TABLE->M_RECORD_BUFFER.RECORD_SIZE() ==               RECORD_PREFIX_SIZE(TAB)' FAIL
Index merge scans might need to read the primary key columns in
addition to the columns that are in the read set. When a record buffer
is allocated for an index merge, space is allocated for the columns in
the read set and also for the primary key columns if needed. For a
dynamic range scan, however, the extra space for the primary key
columns is not allocated unless the first range is read using an index
merge. If the dynamic range scan later uses an index merge to read a
range, it might not have enough space for the primary key columns in
the record buffer.

This is fixed by allocating space for the primary key columns in the
record buffer also for dynamic range scans.

Change-Id: I365cd9e26f96f64f5dbea30f62cce86d4723ad3c