Bug #24945638 STOPONERROR = 0 WITH UNCONTROLLED EXIT RESTARTS IN SAME WAY AS PREVIOUS RESTART
On ndbmtd crash/shutdown, the angel sets the restart flags to
Initial or Normal, and to Start or Nostart, depending upon
parameters like the type of shutdown. For the case where the
shutdown type is an uncontrolled exit and the StopOnError config
parameter is set to false, the restart flags were not set, and
defaulted to Initial/Start. If all the ndbmtds had uncontrolled
crashes within a specific time interval, then all the ndbmtds were
restarted in Initial mode and all the data was lost. Added a fix
for the case of StopOnError=false and uncontrolled exit, to set
the flags to Normal/Start so that ndbmtd does not lose data.

In the case where the ndbmtd crashes so hard that it cannot send
startphase information to the angel, the crash is always treated
as a startup failure. Added a check to execute startup failure
handling only if a valid startphase has been received from the
client.