WL#10196 Dbacc: Remove static array of 8K pages.
8K pages are allocated from 32K pages.

32K pages are seized and released from data memory as needed.

Note, that 32K pages only uses 30-bit physical page number since two bits
are needed to address one 8K sub page.

The logic for keep percentage of data memory free are removed, letting
the general logic in Cmvmi handle that.

Expand and shrink of the hash table are allowed to use the free
percentage of spare pages of data memory.

The page map from logical to physical page number are changed to use
27 bit physical page numbers instead of 19 bits as earlier.