Bug #25358534: REMOVE UNNEEDED ZERO EXTENSIONS FROM CRC32 CODE
ut_crc32() has a number of unneeded zero extensions:

      66:       f2 48 0f 38 f1 42 80    crc32q -0x80(%rdx),%rax
      6d:       89 c0                   mov    %eax,%eax
      6f:       f2 48 0f 38 f1 42 88    crc32q -0x78(%rdx),%rax
      76:       89 c0                   mov    %eax,%eax
      78:       f2 48 0f 38 f1 42 90    crc32q -0x70(%rdx),%rax
      7f:       89 c0                   mov    %eax,%eax
      81:       f2 48 0f 38 f1 42 98    crc32q -0x68(%rdx),%rax
      88:       89 c0                   mov    %eax,%eax

(etc.)

These are completely unneeded, as the crc32q extension zero-extends by itself
even when writing to a 64-bit register. Simply change to a 64-bit type everywhere,
which removes the zero extensions.

The microbenchmarks were written to work with the new microbenchmark framework,
which gives much more stable timings. Results (64-bit, opt, Skylake 3.4 GHz):

  BM_CRC32             2408 -> 1787 ns/iter [+34.8%]
  BM_BigEndianCRC32    2419 -> 1807 ns/iter [+33.9%]

which is something like 0.5% in a typical sysbench run (the function goes
down from ~3.8% to ~3.2% of CPU time used). This matches well up with theory,
as crc32q has 3 cycles latency and mov is a bit under 1, and they are mutually
dependent.

Change-Id: I22f16d1a9cce98bcfd74e10d19ab69566b655e13