Bug#25339088 REGRESSION IN SYSBENCH/POINT_SELECT DUE TO CTE WL
The WL added 4 integers and one pointer to TABLE_SHARE (one of those
integers was actually moved from TABLE). This caused a repeatable 4% drop
in performance in sysbench's point_select test.
Fix: replace those 5 members with a single pointer to a separate
structure specific ot internal temporary tables; other tables
logically shouldn't reserve room for useless information.
This fully regains the lost performance.
As the struct is specific of tmp tables, removed the "tmp" word
from members' names.
Do initialization of the structure in init_tmp_table_share; adding a
new argument to that latter function, with default NULL; using that
opportunity to make the previous argument (MEM_ROOT*), added by the
CTE WL, have a similar default, so most calls become shorter (as they
were before the CTE WL).
Limiting to max 1024 references to a single CTE so that we can use a
short type (uint16) for handler_count to have a small struct (and this
limit makes sense anyway). Adding error message and test of this limit.