Bug #24671968: WHEN THE OPTIMISER IS USING INDEX FOR                GROUP-BY IT OFTEN GIVES WRONG RESULTS
Issue:
------
This problem occurs under the conditions:
1) WHERE clause contains a dependent subquery.
2) Table has a secondary index on the columns in the SELECT
   list, followed by the columns in the subquery.
3) Group-by / DISTINCT allows the query to pick loose index
   scan.

Based on conditions 1 and 2 optimizer chooses not to create
range tree for the WHERE clause due to the dependent
subquery.

This effectively means that without the range tree, the
function evaluating the possibility of loose index scan
will assume that the WHERE clause doesn't contain anything
that would disqualify it from using this approach. Hence
the loose index scan structures are created by taking only
the SELECT list into account. This gives incorrect results
since a smaller prefix of the index is chosen for loose
index scan.

Solution:
---------
When there is a WHERE condition without a range tree, any
reference to the columns in the secondary index that are so
far unused, will disqualify the query from using loose
index scan. But this check is not performed for subqueries
and this has been corrected.