BUG#11763201 PLEASE REMOVE CLIENT PROGRAM VERSIONS
Problem: a number of mysql clients and utilities report a version
string when for instance started with the '-V' option. The version
numbers reported aren't consistently updated and don't make sense any
more. Also, the version strings vary with respect to punctuation and
use of parentheses. This looks untidy and not very professional.

Solution: keep the version string reported by the server, and make
that the standard for all clients/utilities. Do not report specific
version numbers.

The patch also adds the standard Oracle welcome notice to a few
clients that didn't use it.

Approved by Jon Olav Hauglid <jon.hauglid@oracle.com>