Bug #25092498: NDB API PROCESS CRASH (CHECKMAGICNUMBER (THIS=0X0, B=FALSE) AT /PLATSOFTWARE/MYS
checkMagicNumber() fails due to a NULL pointer being passed to it.
This happens due to simultaneous access of theNdbObjectIdMap without
proper synchronisation/protection.

The NdbObjectIdMap object is internally expanded using realloc() when needed.
realloc() sometimes changes the memory location of the internal map object
and the old memory object is freed.
Hence, when one thread causes the map to expand, another thread might
still be accessing the old location of the map. This could cause a
segmentation fault in the thread accessing the old location.

This patch adds mutex protection to theNdbObjectIdMap by reusing the
existing mutex in NdbImpl: m_mutex.
This mutex is already being locked before reading theNdbObjectIdMap.
Now, the mutex is locked before theNdbObjectIdMap is expanded as well.
We don't need to lock it before unmap() since it's guaranteed to be
called only after map()/expand().