BUG#25311164 - UPDATE XCOM ON GROUP REPLICATION
This updates GR to the latest XCom on GCS. On this commit we have fixes for:

BUG#22204121 - NODE THAT FAILS AND RE-ENTERS MIGHT BE EXPELLED WITH NO REASON
Added the notion of reincarnation of a node.

BUG#22812902 - ENABLE WERROR AS DEFAULT IN MYSQL-GCS BUILDS
Several warning corrections

BUG#22671683 - LIMIT MEMORY UTILIZATION IN XCOM
Adding cache methods to avoid breaking examples code
Added minimum delivered message number to xcom messages to avoid deallocating cache that is needed by slow or joining nodes. Incread xcom protocol version to x_1_2
Added functions to enforce an upper limit to the application data in the xcom cache

Some changes towards this adaptation:
- Corrected Copyright year
- Corrected invalid declaration
- Adapting to printing size_t on 32-bit platforms
- Removing unecessary declarations
- Corrected the fact that the patch was not self-sufficient
  in terms of not using UUIDs with XCom