Bug#25372667: MAIN.OPT_HINTS_PFS FAILS ON PB2 OCCASIONALLY
Issue:
------
main.opt_hints_pfs has been failing on pb2 sporadically
with an error saying that the maximum statement execution
time exceeded for a query that uses the MAX_EXECUTION_TIME
hint.

Fix:
----
The maximum statement execution time is exceeded only once in a
while if pb2 is loaded. When it is loaded, the response time
increases and determining timeout values in tests can be very hard.
The fix is to set the timeout to a large value.

Reviewed-by: Gleb Shchepa <gleb.shchepa@oracle.com>
RB: 15061